+++ deeppavlov/configs/go_bot/gobot_dstc2_rule_base.json
