+++ deeppavlov/configs/go_bot/gobot_dstc2_simple.json
